core/requirements.txt
jira==3.5.1
openai==1.3.5
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0
